Make resource_plugin_config more generic. Remove data source for plugin config since it does not provide any  value. Fix tests.
