+++ bitbucket/provider.go
@@ -38,7 +38,6 @@ func Provider() terraform.ResourceProvider {
 			"bitbucketserver_groups":                        dataSourceGroups(),
 			"bitbucketserver_group_users":                   dataSourceGroupUsers(),
 			"bitbucketserver_plugin":                        dataSourcePlugin(),
-			"bitbucketserver_plugin_config":                 dataSourcePluginConfig(),
 			"bitbucketserver_project_hooks":                 dataSourceProjectHooks(),
 			"bitbucketserver_project_permissions_groups":    dataSourceProjectPermissionsGroups(),
 			"bitbucketserver_project_permissions_users":     dataSourceProjectPermissionsUsers(),
